pharmacy_portal/requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
bcrypt==4.0.1
pandas==2.0.3
openpyxl==3.1.2
Werkzeug==2.3.7
twilio==8.2.0
python-dotenv==1.0.0
